fix: moved academy folder in buildings
